[mod] Moves all dialogs to separate folder
