+++ src/components/annotator/AnnotatorDrawer/OpenMenu/OpenImageMenuItem.tsx
@@ -4,7 +4,7 @@ import { ListItemText, MenuItem } from "@mui/material";
 
 import { useUpload } from "hooks";
 
-import { ImageShapeDialog } from "components/common/dialogs";
+import { ImageShapeDialog } from "components/dialogs";
 import { ImageShapeEnum, ImageShapeInfo } from "utils/common/image";
 
 type OpenImageMenuItemProps = {
