--- src/components/common/dialogs/index.ts
@@ -1,10 +1,0 @@
-export { AlertDialog } from "./AlertDialog";
-export { CustomDialog } from "./CustomDialog";
-export { DialogWithAction } from "./DialogWithAction";
-export { DialogTransitionSlide } from "./DialogTransitionSlide";
-export { FallBackDialog } from "./FallBackDialog";
-export { ImageShapeDialog } from "./ImageShapeDialog";
-export { ImportTensorflowModelDialog } from "./ImportTensorflowModelDialog/ImportTensorflowModelDialog";
-export { SendFeedbackDialog } from "./SendFeedbackDialog";
-export { SettingsDialog } from "./SettingsDialog";
-export { UpsertCategoriesDialog } from "./UpsertCategoriesDialog";
