+++ src/components/common/list-items/SendFeedbackListItem/SendFeedbackListItem.tsx
@@ -6,7 +6,7 @@ import FeedbackIcon from "@mui/icons-material/Feedback";
 
 import { useDialog } from "hooks";
 
-import { SendFeedbackDialog } from "../../dialogs";
+import { SendFeedbackDialog } from "../../../dialogs";
 
 export const SendFeedbackListItem = () => {
   const { onClose, onOpen, open } = useDialog();
