+++ src/components/common/list-items/SettingsListItem/SettingsListItem.tsx
@@ -5,7 +5,7 @@ import SettingsIcon from "@mui/icons-material/Settings";
 
 import { useDialog } from "hooks";
 
-import { SettingsDialog } from "../../dialogs";
+import { SettingsDialog } from "../../../dialogs";
 
 export const SettingsListItem = () => {
   const { onClose, onOpen, open } = useDialog();
