+++ src/components/dialogs/ExampleImageDialog/ExampleImageDialog.tsx
