+++ src/components/dialogs/ExampleImageDialog/OpenExampleImageMenuItem.tsx
