+++ src/components/dialogs/ExampleImageDialog/index.ts
