+++ src/components/main/MainAppBar/MainAppBar.tsx
@@ -5,7 +5,7 @@ import { AppBar, Box } from "@mui/material";
 
 import { MainToolbar } from "../MainToolbar";
 
-import { AlertDialog } from "components/common/dialogs/";
+import { AlertDialog } from "components/dialogs";
 
 import { alertStateSelector } from "store/application";
 import { APPLICATION_COLORS } from "utils/common/colorPalette";
