+++ src/components/segmenter/SegmenterListItems/FitSegmenterListItem.tsx
@@ -9,7 +9,7 @@ import { ScatterPlot as ScatterPlotIcon } from "@mui/icons-material";
 
 import { useDialog } from "hooks";
 
-import { FitSegmenterDialog } from "../FitSegmenterDialog";
+import { FitSegmenterDialog } from "components/dialogs";
 
 export const FitSegmenterListItem = () => {
   const { onClose, onOpen, open } = useDialog(false);
